Fix address sanitizer build option
The address sanitizer flags were being set added after the rocsolver
library target was created, which meant they only applied to later
targets. Using an interface library is a little more annoying, but
prevents this problem.

The clients need additional flags to link with rocSOLVER when
address sanitizer is enabled, so disable them for now in sanitizer
builds.

Drop stray link against fmt::fmt, which was leftover from a bad merge.
(The header-only option should take precidence anyway, so this is
mostly unimportant.)