Convert to ReStructured Text
rocBLAS deduplicates the contributor's guide by creating
`docs/source/Contributors_Guide.rst` as a symlink to `CONTRIBUTING.rst`

That solution is not necessarily ideal, but there is value in being
able to share text between the documentation and the contributing file
without requiring conversion. Using ReStructured Text now will make it
easier to improve upon this in the future.