Minimize root privilages during deps build
The Problem:

When installing client dependencies with ./install.sh -cd, the new git
safedirectory restriction will cause the CMake ExternalProject_Add to
fail. Prior to this change, the actual GTest and LAPACK builds were
done during the `sudo make install` step, so the user for the build was
being done as root, but the GTest and LAPACK repositories were owned by
the user that invoked install.sh. It seems that ExternalProject_Add
would query the status of external git projects it managed, which would
fail as the repo would not be owned by root (unless install.sh was
invoked as root).

The Fix:

Only use root privilages when installing the deps, not when building
them. This is a port of Torre's fix for rocBLAS in
https://github.com/ROCmSoftwarePlatform/rocBLAS/commit/46b04c6